etl/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.1
elasticsearch==8.10.0
numpy==1.26.0
python-dotenv==1.0.0
